UI: Refactor duplicated template code into a common render() helper
We remove the calls to setCharacterSet(UTF_8) since
WriterRepresentation's constructor does that anyway.
