+++ engine/src/main/java/org/archive/crawler/restlet/BaseResource.java
@@ -19,7 +19,18 @@
 
 package org.archive.crawler.restlet;
 
+import freemarker.template.ObjectWrapper;
+import freemarker.template.Template;
+import freemarker.template.TemplateException;
+import org.archive.crawler.restlet.models.ViewModel;
+import org.restlet.data.MediaType;
+import org.restlet.representation.Representation;
+import org.restlet.representation.WriterRepresentation;
 import org.restlet.resource.ServerResource;
+
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.io.Writer;
 
 /**
  * Abstract {@code Resource} with common shared functionality. 
@@ -35,5 +46,37 @@ public abstract class BaseResource extends ServerResource {
     protected String getStaticRef(String resource) {
         String rootRef = getRequest().getRootRef().toString();
         return rootRef + "/engine/static/" + resource;
+    }
+
+    protected Representation render(String templateName, ViewModel viewModel) {
+        return render(templateName, viewModel, null);
+    }
+
+    protected Representation render(String templateName, ViewModel viewModel, ObjectWrapper objectWrapper) {
+        String baseRef = getRequest().getResourceRef().getBaseRef().toString();
+        if(!baseRef.endsWith("/")) {
+            baseRef += "/";
+        }
+        viewModel.put("baseRef", baseRef);
+        viewModel.setFlashes(Flash.getFlashes(getRequest()));
+
+        Template template;
+        try {
+            template = getApplication().getTemplateConfiguration().getTemplate(templateName);
+        } catch (IOException e) {
+            throw new UncheckedIOException("Error reading template " + templateName, e);
+        }
+
+        return new WriterRepresentation(MediaType.TEXT_HTML) {
+            @Override
+            public void write(Writer writer) throws IOException {
+                try {
+                    template.process(viewModel, writer, objectWrapper);
+                } catch (TemplateException e) {
+                    throw new RuntimeException(e);
+                }
+                writer.flush();
+            }
+        };
     }
 }
