+++ engine/src/main/java/org/archive/crawler/restlet/EngineResource.java
@@ -25,14 +25,13 @@ import java.io.Writer;
 import java.util.HashMap;
 import java.util.Map;
 
+import freemarker.template.ObjectWrapper;
 import org.archive.crawler.framework.CrawlJob;
-import org.archive.crawler.framework.Engine;
 import org.archive.crawler.restlet.models.EngineModel;
 import org.archive.crawler.restlet.models.ViewModel;
 import org.restlet.Context;
 import org.restlet.Request;
 import org.restlet.Response;
-import org.restlet.data.CharacterSet;
 import org.restlet.data.Form;
 import org.restlet.data.MediaType;
 import org.restlet.representation.EmptyRepresentation;
@@ -40,11 +39,6 @@ import org.restlet.representation.Representation;
 import org.restlet.representation.WriterRepresentation;
 import org.restlet.resource.ResourceException;
 import org.restlet.representation.Variant;
-
-import freemarker.template.Configuration;
-import freemarker.template.DefaultObjectWrapper;
-import freemarker.template.Template;
-import freemarker.template.TemplateException;
 
 import static org.restlet.data.MediaType.APPLICATION_XML;
 
@@ -67,23 +61,18 @@ public class EngineResource extends BaseResource {
 
     @Override
     protected Representation get(Variant variant) throws ResourceException {
-        Representation representation;
         if (variant.getMediaType() == APPLICATION_XML) {
-            representation = new WriterRepresentation(APPLICATION_XML) {
+            return new WriterRepresentation(APPLICATION_XML) {
                 public void write(Writer writer) throws IOException {
                     XmlMarshaller.marshalDocument(writer, "engine", makeDataModel());
                 }
             };
         } else {
-            representation = new WriterRepresentation(MediaType.TEXT_HTML) {
-                public void write(Writer writer) throws IOException {
-                    EngineResource.this.writeHtml(writer);
-                }
-            };
+            ViewModel viewModel = new ViewModel();
+            viewModel.put("fileSeparator", File.separator);
+            viewModel.put("engine", makeDataModel());
+            return render("Engine.ftl", viewModel, ObjectWrapper.DEFAULT_WRAPPER);
         }
-        // TODO: remove if not necessary in future?
-        representation.setCharacterSet(CharacterSet.UTF_8);
-        return representation;
     }
 
     @Override
@@ -211,31 +200,6 @@ public class EngineResource extends BaseResource {
         }
         
         return new EngineModel(getEngine(), baseRef);
-    }
-    
-    protected void writeHtml(Writer writer) {
-        EngineModel model = makeDataModel();
-        String baseRef = getRequest().getResourceRef().getBaseRef().toString();
-        if(!baseRef.endsWith("/")) {
-            baseRef += "/";
-        }
-        Configuration tmpltCfg = getApplication().getTemplateConfiguration();
-
-        ViewModel viewModel = new ViewModel();
-        viewModel.setFlashes(Flash.getFlashes(getRequest()));
-        viewModel.put("baseRef",baseRef);
-        viewModel.put("fileSeparator", File.separator);
-        viewModel.put("engine", model);
-
-        try {
-            Template template = tmpltCfg.getTemplate("Engine.ftl");
-            template.process(viewModel, writer);
-            writer.flush();
-        } catch (IOException e) { 
-            throw new RuntimeException(e); 
-        } catch (TemplateException e) { 
-            throw new RuntimeException(e); 
-        }
     }
 
     protected Engine getEngine() {
