+++ engine/src/main/java/org/archive/crawler/restlet/ScriptResource.java
@@ -19,39 +19,26 @@
 
 package org.archive.crawler.restlet;
 
-import java.io.IOException;
-import java.io.Writer;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
-import javax.script.ScriptEngine;
-import javax.script.ScriptEngineFactory;
-import javax.script.ScriptEngineManager;
-
 import org.apache.commons.lang.StringUtils;
 import org.archive.crawler.restlet.models.ScriptModel;
 import org.archive.crawler.restlet.models.ViewModel;
 import org.restlet.Context;
-import org.restlet.data.CharacterSet;
+import org.restlet.Request;
+import org.restlet.Response;
 import org.restlet.data.Form;
 import org.restlet.data.MediaType;
 import org.restlet.data.Reference;
-import org.restlet.Request;
-import org.restlet.Response;
 import org.restlet.representation.Representation;
+import org.restlet.representation.Variant;
 import org.restlet.representation.WriterRepresentation;
 import org.restlet.resource.ResourceException;
-import org.restlet.representation.Variant;
 
-import freemarker.template.Configuration;
-import freemarker.template.ObjectWrapper;
-import freemarker.template.Template;
-import freemarker.template.TemplateException;
+import javax.script.ScriptEngine;
+import javax.script.ScriptEngineFactory;
+import javax.script.ScriptEngineManager;
+import java.io.IOException;
+import java.io.Writer;
+import java.util.*;
 
 /**
  * Restlet Resource which runs an arbitrary script, which is supplied
@@ -114,23 +101,20 @@ public class ScriptResource extends JobRelatedResource {
 
     @Override
     public Representation get(Variant variant) throws ResourceException {
-        Representation representation;
         if (variant.getMediaType() == MediaType.APPLICATION_XML) {
-            representation = new WriterRepresentation(MediaType.APPLICATION_XML) {
+            return new WriterRepresentation(MediaType.APPLICATION_XML) {
                 public void write(Writer writer) throws IOException {
                     XmlMarshaller.marshalDocument(writer,"script", makeDataModel());
                 }
             };
         } else {
-            representation = new WriterRepresentation(MediaType.TEXT_HTML) {
-                public void write(Writer writer) throws IOException {
-                    ScriptResource.this.writeHtml(writer);
-                }
-            };
+            ViewModel viewModel = new ViewModel();
+            viewModel.put("baseResourceRef", getRequest().getRootRef().toString() + "/engine/static/");
+            viewModel.put("model", makeDataModel());
+            viewModel.put("selectedEngine", chosenEngine);
+            viewModel.put("staticRef", getStaticRef(""));
+            return render("Script.ftl", viewModel);
         }
-        // TODO: remove if not necessary in future?
-        representation.setCharacterSet(CharacterSet.UTF_8);
-        return representation;
     }
 
     protected Collection<Map<String,String>> getAvailableScriptEngines() {
@@ -164,33 +148,5 @@ public class ScriptResource extends JobRelatedResource {
                 getAvailableScriptEngines());
 
         return model;
-    }
-
-    protected void writeHtml(Writer writer) {
-        
-        String baseRef = getRequest().getResourceRef().getBaseRef().toString();
-        if(!baseRef.endsWith("/")) {
-            baseRef += "/";
-        }
-        Configuration tmpltCfg = getApplication().getTemplateConfiguration();
-
-        ViewModel viewModel = new ViewModel();
-        viewModel.setFlashes(Flash.getFlashes(getRequest()));
-        viewModel.put("baseRef",baseRef);
-        viewModel.put("staticRef", getStaticRef(""));
-        viewModel.put("baseResourceRef",getRequest().getRootRef().toString()+"/engine/static/");
-        viewModel.put("model", makeDataModel());
-        viewModel.put("selectedEngine", chosenEngine);
-
-        try {
-            Template template = tmpltCfg.getTemplate("Script.ftl");
-            template.process(viewModel, writer);
-            writer.flush();
-        } catch (IOException e) { 
-            throw new RuntimeException(e); 
-        } catch (TemplateException e) { 
-            throw new RuntimeException(e); 
-        }
-
     }
 }
